+++ src/compiler/typescript_to_rust/azle_generate/src/errors/mod.rs
@@ -1,112 +1,7 @@
-use annotate_snippets::{
-    display_list::{DisplayList, FormatOptions},
-    snippet::{Annotation, AnnotationType, Slice, Snippet, SourceAnnotation},
-};
-use std::fmt;
+mod error_message;
+mod suggestion;
+
+pub use error_message::ErrorMessage;
+pub use suggestion::Suggestion;
 
 pub mod external_canister_method;
-
-pub struct Suggestion {
-    pub title: String,
-    pub source: String,
-    pub range: (usize, usize),
-    pub annotation: Option<String>,
-    pub import_suggestion: Option<String>,
-}
-
-pub struct ErrorMessage {
-    pub title: String,
-    pub origin: String,
-    pub line_number: usize,
-    pub source: String,
-    pub range: (usize, usize),
-    pub annotation: String,
-    pub suggestion: Option<Suggestion>,
-}
-
-impl ErrorMessage {
-    fn to_string(&self) -> String {
-        let error_snippet = Snippet {
-            title: Some(Annotation {
-                label: Some(&self.title),
-                id: None,
-                annotation_type: AnnotationType::Error,
-            }),
-            footer: vec![],
-            slices: vec![Slice {
-                source: &self.source,
-                line_start: self.line_number,
-                origin: Some(&self.origin),
-                fold: true,
-                annotations: vec![SourceAnnotation {
-                    label: &self.annotation,
-                    annotation_type: AnnotationType::Error,
-                    range: self.range,
-                }],
-            }],
-            opt: FormatOptions {
-                color: true,
-                ..Default::default()
-            },
-        };
-
-        match &self.suggestion {
-            None => format!("{}", DisplayList::from(error_snippet)),
-            Some(suggestion) => {
-                let suggestion_slice = Slice {
-                    source: &suggestion.source,
-                    line_start: self.line_number,
-                    origin: None,
-                    fold: false,
-                    annotations: vec![SourceAnnotation {
-                        label: match &suggestion.annotation {
-                            Some(annotation) => &annotation,
-                            None => "",
-                        },
-                        annotation_type: AnnotationType::Help,
-                        range: suggestion.range,
-                    }],
-                };
-                let slices = match &suggestion.import_suggestion {
-                    Some(import) => vec![
-                        Slice {
-                            source: &import,
-                            line_start: 1,
-                            origin: None,
-                            annotations: vec![],
-                            fold: false,
-                        },
-                        suggestion_slice,
-                    ],
-                    None => vec![suggestion_slice],
-                };
-
-                let suggestion_snippet = Snippet {
-                    title: Some(Annotation {
-                        label: Some(&suggestion.title),
-                        id: None,
-                        annotation_type: AnnotationType::Help,
-                    }),
-                    footer: vec![],
-                    slices,
-                    opt: FormatOptions {
-                        color: true,
-                        ..Default::default()
-                    },
-                };
-
-                format!(
-                    "{}\n{}",
-                    DisplayList::from(error_snippet),
-                    DisplayList::from(suggestion_snippet)
-                )
-            }
-        }
-    }
-}
-
-impl fmt::Display for ErrorMessage {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", self.to_string())
-    }
-}
