+++ src/compiler/typescript_to_rust/azle_generate/src/errors/suggestion.rs
@@ -1,0 +1,7 @@
+pub struct Suggestion {
+    pub title: String,
+    pub source: String,
+    pub range: (usize, usize),
+    pub annotation: Option<String>,
+    pub import_suggestion: Option<String>,
+}
